+++ src/pages/OnboardingPage/StepStart/StepStart.tsx
