--- src/pages/OnboardingPage/components/index.ts
@@ -1,3 +1,0 @@
-export { default as StepStart } from './Step/StepStart/StepStart.tsx';
-export { default as StepField } from './Step/StepField/StepField.tsx';
-export { default as StepService } from './Step/StepService/StepService.tsx';
